// src/module_bindings/types.ts
// THIS FILE IS AUTOMATICALLY GENERATED BY SPACETIMEDB. EDITS TO THIS FILE
// WILL NOT BE SAVED. MODIFY TABLES IN YOUR MODULE SOURCE CODE INSTEAD.

/* eslint-disable */
/* tslint:disable */
import {
  TypeBuilder as __TypeBuilder,
  t as __t,
  type AlgebraicTypeType as __AlgebraicTypeType,
  type Infer as __Infer,
} from "spacetimedb";

export const Floor = __t.object("Floor", {
  id: __t.string(),
  hosting: __t.bool(),
  scriptIndex: __t.u32(),
  lastScriptAtMs: __t.i64(),
  nextId: __t.u32(),
  toastId: __t.option(__t.string()),
  toastTitle: __t.option(__t.string()),
  toastBody: __t.option(__t.string()),
});
export type Floor = __Infer<typeof Floor>;

export const Ingredient = __t.object("Ingredient", {
  id: __t.string(),
  name: __t.string(),
  quantityLabel: __t.string(),
  grams: __t.f64(),
  category: __t.string(),
  priceUsd: __t.f64(),
  expiresAtMs: __t.i64(),
  addedAtMs: __t.i64(),
  source: __t.string(),
  state: __t.string(),
  note: __t.option(__t.string()),
  safety: __t.option(__t.string()),
  fromName: __t.option(__t.string()),
  eventId: __t.option(__t.string()),
});
export type Ingredient = __Infer<typeof Ingredient>;

export const RecipeRun = __t.object("RecipeRun", {
  id: __t.string(),
  atMs: __t.i64(),
  recipeId: __t.string(),
  recipeName: __t.string(),
  ingredientIds: __t.string(),
});
export type RecipeRun = __Infer<typeof RecipeRun>;

export const RescuePost = __t.object("RescuePost", {
  id: __t.string(),
  ingredientId: __t.option(__t.string()),
  name: __t.string(),
  quantityLabel: __t.string(),
  grams: __t.f64(),
  category: __t.string(),
  priceUsd: __t.f64(),
  expiresAtMs: __t.i64(),
  poster: __t.string(),
  postedAtMs: __t.i64(),
  note: __t.string(),
  safety: __t.option(__t.string()),
  claimedBy: __t.option(__t.string()),
  claimedAtMs: __t.option(__t.i64()),
});
export type RescuePost = __Infer<typeof RescuePost>;

export const RoomNote = __t.object("RoomNote", {
  id: __t.string(),
  atMs: __t.i64(),
  author: __t.string(),
  text: __t.string(),
});
export type RoomNote = __Infer<typeof RoomNote>;

export const WasteEvent = __t.object("WasteEvent", {
  id: __t.string(),
  atMs: __t.i64(),
  kind: __t.string(),
  name: __t.string(),
  grams: __t.f64(),
  category: __t.string(),
  priceUsd: __t.f64(),
  co2EKg: __t.f64(),
  methaneEmittedKg: __t.f64(),
  embodiedCo2EKg: __t.f64(),
  waterL: __t.f64(),
});
export type WasteEvent = __Infer<typeof WasteEvent>;